Run Babel at build time instead of runtime and produce standalone ES5 scripts.
Run Babel as part of the build instead of a runtime require hook. There is no
runtime dependency on babel-core anymore and the scripts can now run without
requiring babel-register and a runtime require hook.

Move source files to src directories, and point package.json to Babel's ES5
output in lib directories, to separate the ES6 sources from the Babel output.

Adjust the build tools to the new src and lib locations and the removal of the
runtime dependency on Babel-core.

Add the necessary .npmrc and .gitignore files to all the packages.